Separate out icon list and Component wrapper for svgs
