+++ libs/ui/src/lib/icon/__stories__/Icon.stories.mdx
@@ -4,7 +4,8 @@ import { ArgsTable, Canvas, Story, Meta } from '@storybook/addon-docs/blocks'
 import styled from 'styled-components'
 import { colorPalette } from '@oxide/theme'
 
-import { Icon, icons } from '../Icon'
+import { Icon } from '../Icon'
+import { icons } from '../icons'
 import { Default, Rotate, CustomTitle } from './Icon.stories'
 
 <Meta
